+++ src/frontend/src/pages/Auth/Register.tsx
@@ -3,7 +3,7 @@ import { Anchor, Text } from '@mantine/core';
 import { useNavigate } from 'react-router-dom';
 import { RegistrationForm } from '../../components/forms/AuthenticationForm';
 import {} from '../../functions/auth';
-import { Wrapper } from './LoginLayoutComponent';
+import { Wrapper } from './Layout';
 
 export default function Register() {
   const navigate = useNavigate();
